pyproject.toml
[tool.poetry]
name = "evosim"
version = "0.1.0"
description = ""
authors = ["Nathaniel Been"]
readme = "README.md"

[tool.poetry.scripts]
evosim = "main:cli"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.1.0"
opencv-python = "^4.8.1.78"
numpy = "^1.26.2"
pygraphviz = "^1.11"
matplotlib = "^3.8.2"
click = "^8.1.7"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
